feat: move collection UI types + pure frontend utils into the package (chunk 1d step 1)
Move the UI-only view-state types (EditState, TableRowDraft, Ref*/Embed caches,
RefOption, CollectionRecordIssue, CollectionDetailResponse, ItemMutationResponse)
into the package core (core/uiTypes.ts), and the pure frontend utils enumColors
(enum→Tailwind class mapping) and draft (form draft↔record coercion) into the
core. The host collectionTypes.ts now re-exports every collection type from the
package — it owns none.

The Collection* components still reach enumColors/draft through thin re-export
shims at the old util paths; those shims go away in a later step when the
components themselves move into the package.

First step toward the View extraction (no host-capabilities injection needed —
these are pure). Verified: yarn typecheck (server/vue/test/packages), lint,
build all green; full unit suite (5102 tests) 0 failures.
